Changed TotalSpeakers variable to Double

+++ src/main/java/com/napier/sem/Language.java
@@ -2,6 +2,6 @@ package com.napier.sem;
 
 public class Language {
     public String Language;
-    public long TotalSpeakers;
+    public double TotalSpeakers;
     public double PercentOfWorldPop;
 }
